server crud with auth

+++ components/button.tsx
@@ -2,6 +2,7 @@
 
 import { deleteUser } from "@/data/route";
 import { toast } from "react-toastify";
+import Swal from "sweetalert2";
 
 interface Props {
   id: number;
@@ -9,9 +10,34 @@ interface Props {
 
 const ButtonTest: React.FC<Props> = ({ id }) => {
   const destroy = async () => {
-    const res = await deleteUser(id);
-    console.log(res);
-    
+    try {
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          const response = await deleteUser(id);
+          if (response.errors) {
+            const errorMessages = Object.values(response.errors).flat();
+            toast.error(errorMessages.join(", "));
+          } else {
+            toast.success(response.message);
+          }
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your file has been deleted.",
+            icon: "success",
+          });
+        }
+      });
+    } catch (error) {
+      console.error("Error submitting form:", error);
+    }
   };
 
   return (
